requirements.txt
django
gunicorn
daphne
psycopg2
markdown
django-filter 
graphene-django
django-crispy-forms
crispy-bootstrap5
django-bulk-update
coverage
django-debug-toolbar
channels
Twisted[tls,http2]
hiredis
channels-redis
celery
django-celery-beat
django-environ
bleach
cryptography
pandas
numpy
grequests
alive-progress
requests
django-phone-field
django-countries
Pillow